requirements.txt
telethon
openai
sqlite-utils
python-dotenv
